Do a Warp instead of re-initializing the same zone
See #141

+++ src/ipc/zone/server/warp.rs
@@ -1,11 +1,13 @@
 use binrw::binrw;
 
-use crate::common::Position;
+use crate::common::{Position, read_quantized_rotation, write_quantized_rotation};
 
 #[binrw]
 #[derive(Debug, Clone, Default)]
 pub struct Warp {
-    pub dir: u16,
+    #[br(map = read_quantized_rotation)]
+    #[bw(map = write_quantized_rotation)]
+    pub dir: f32,
     pub warp_type: u8,
     pub warp_type_arg: u8,
     pub layer_set: u32,
